+++ example/get_spiner_bounds.cpp
@@ -29,25 +29,30 @@ int main(int argc, char *argv[]) {
 #endif
   { // begin kokkos scope
     if (argc < 3) {
-      std::cerr << "Usage: " << argv[0] << " matid sp5_filename" << std::endl;
-      // std::cerr << "Arguments given: " << argc << std::endl;
+      std::cerr << "Usage: " << argv[0] << " matid sp5_filename [avoid vapor dome]"
+                << std::endl;
       std::exit(1);
     }
-
     const int matid = std::atoi(argv[1]);
     const std::string filename = argv[2];
+    bool pmin_vapor_dome = false;
+    if (argc >= 4) {
+      pmin_vapor_dome = atoi(argv[3]);
+    }
+    singularity::SpinerEOSDependsRhoT deprt(filename, matid, false, pmin_vapor_dome);
+    singularity::SpinerEOSDependsRhoSie depre(filename, matid, false, pmin_vapor_dome);
 
-    singularity::SpinerEOSDependsRhoT deprt(filename, matid);
-    singularity::SpinerEOSDependsRhoSie depre(filename, matid);
-
-    printf("rho bounds, depends(r, T): %.14e %.14e\n", deprt.MinimumDensity(),
+    printf("# rho bounds, depends(r, T): %.14e %.14e\n", deprt.MinimumDensity(),
            deprt.rhoMax());
-    printf("rho bounds, depends(r, sie): %.14e %.14e\n", depre.MinimumDensity(),
+    printf("# rho bounds, depends(r, sie): %.14e %.14e\n", depre.MinimumDensity(),
            depre.rhoMax());
-    printf("T bounds, depends(r, T): %.14e %.14e\n", deprt.MinimumTemperature(),
+    printf("# T bounds, depends(r, T): %.14e %.14e\n", deprt.MinimumTemperature(),
            deprt.TMax());
-    printf("T bounds, depends(r, sie): %.14e %.14e\n", depre.MinimumTemperature(),
+    printf("# T bounds, depends(r, sie): %.14e %.14e\n", depre.MinimumTemperature(),
            depre.TMax());
+    printf("\n");
+    printf("# [0]: iT, [1]: T, [2] rho(Pmin, T)\n");
+    deprt.PrintRhoPMin();
 
     deprt.Finalize();
     depre.Finalize();