+++ test/test_closure_pte.cpp
@@ -417,6 +417,13 @@ SCENARIO("Density- and Pressure-Temperature PTE Solvers", "[PTE]") {
     EOS foam_eos_h = SpinerEOSDependsRhoT(eos_file, foam_matid);
     EOS foam_eos = foam_eos_h.GetOnDevice();
 
+    WHEN("We request density at minimum pressure") {
+      Real rhopmin_100 = foam_eos_h.RhoPmin(100);
+      THEN("It's zero, as appropriate for a Maxwell constructed EOS") {
+        REQUIRE(rhopmin_100 == 0);
+      }
+    }
+
     constexpr Real rhobar1 = 1.59761356859602e-04;
     constexpr Real rhobar2 = 7.81928505957464e-09;
     constexpr Real alpha_guess1 = 9.99999776549350e-01;
@@ -435,6 +442,14 @@ SCENARIO("Density- and Pressure-Temperature PTE Solvers", "[PTE]") {
                      sietot, Tguess, alpha1_true, alpha2_true, Ttrue, Ptrue);
 
       THEN("The solver converges") { REQUIRE(success); }
+
+      AND_WHEN("We call PTE but with a wildly incorrect temperature guess, deep in the "
+               "Maxwell constructed region") {
+        bool success =
+            RunPTE2Mat(He_eos, foam_eos, rhobar1, rhobar2, alpha_guess1, alpha_guess2,
+                       sietot, 5, alpha1_true, alpha2_true, Ttrue, Ptrue);
+        THEN("The solver converges") { REQUIRE(success); }
+      }
     }
 
     He_eos_h.Finalize();