+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - [[PR524]](https://github.com/lanl/singularity-eos/pull/524) Fix future kokkos-kernels compatibility in PTE solver
 
 ### Changed (changing behavior/API/variables/...)
+- [[PR543]](https://github.com/lanl/singularity-eos/pull/543) Made PTE solvers more robust for Maxwell constructed tables
 - [[PR502]](https://github.com/lanl/singularity-eos/pull/502) Expose split tables to Fortran interface
 - [[PR519]] (https://github.com/lanl/singularity-eos/pull/519) Split eos_spiner.hpp into eos_spiner_rho_sie.hpp and eos_spiner_rho_temp.hpp files.